backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.16.1
fsspec==2025.2.0
h11==0.14.0
huggingface-hub==0.29.1
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.5
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
openai-whisper==20240930
packaging==24.2
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
zipp==3.20.2
#triton==3.2.0